requirements.txt
python-dotenv==1.2.1
requests==2.32.5
allure-pytest==2.15.3
pytest==9.0.2
pydantic==2.12.5